Refactor square root calculation to handle negative numbers and improve efficiency

+++ 0x08-recursion/main.h
@@ -8,6 +8,6 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int sqrt_helper(int n, int start, int end);
+long int sqrt_helper(int n, int start, int end);
 
 #endif
